Enforce strict error handling and early aborts
This fixes a number of crashes that currently happen when either the
origin Prometheus cannot be reached or when a user request is faulty.

It also explicitly hard-errors to the user whenever a partial error
happens somewhere in Trickster (due to unreachable origin, faulty JSON
marshaling, or else). There were many cases where an error was logged,
but not handled, causing undefined/crashy behavior. These errors are now
handled and cause the entire request to fail. This seems better than
returning wrong results or crashing.

I also added tests for the obvious crash cases and check whether they
are now returning a user-exposed HTTP error status code instead.

A TODO for the future would be to clean up the metrics handling, since
it currently happens before or after an error return, and errors aren't
properly tracked in the metrics.

Fixes https://github.com/Comcast/trickster/issues/49