Parse the time arg as a float
The time arg in the params can contain decimal points for more
precision. Before this patch that would cause an error on the time
parse, which was ignored, causing the time sent downstream to be `0`.
This patch (1) properly parses the time as a float64-- then casts to
int64 and (2) checks for errors in the parsing