Fix empty response
When comparing to another empty struct this should just be checking if
the pointers match, which they don't. In the past this wasn't an issue
since it had no real bearing on the result-- as the body was parsed out
separately. Now that we are actually using this, we need to check
properly. Instead of using a nil I'm checking that the statuscode is
non-zero.