Validate ordering assumption in mergeMatrix
the current implementation of mergeMatrix assumes that `pe2` is "before"
`pe`-- but does not thing to actually ensure or enforce that this is the
case. This means if there is overlap we can get dupe points. For
example:

```
--- FAIL: TestTricksterHandler_mergeMatrix (0.00s)
    --- FAIL: TestTricksterHandler_mergeMatrix/0 (0.00s)
    	handlers_test.go:542: Mismatch
    		actual={success {matrix a =>
    		1.5 @[0.001]
    		1.5 @[0.002]
    		1.5 @[0.001]}}
    		expected={success {matrix a =>
    		1.5 @[0.001]
    		1.5 @[0.002]}}
```

This is likely the cause for a variety of "dupe points" or other
out-of-order points problems.

Fix #92