Enforce MaxValueAgeSecs in cache lookups
Prior to this patch all requests would do a lookup in the cache and
attempt to "fill" the cache, regardless of age. In addition, responses
from origin were always saved to cache regardless of their window.

A sample scenario to illustrate the point:

1. request shows up for last 3h of data, it is placed in cache
2. requests shows up for 3h of last month's data; this origin request would be
widened to be the last month's data
3. response from origin is "trimmed" to maxValueAgeSecs

With this patch we (1) don't look in the cache for queries referencing
"too old" of data (as we won't have it, and we don't want to fetch the
whole extent). and (2) we don't cache data that isn't within the extent
-- otherwise if we got an old query we'd actually *empty* the cache